requirements.txt
uvicorn==0.31.0
fastapi==0.115.0
python-dotenv==0.20.0
librosa==0.10.2
openai-whisper==20240930
vosk==0.3.45
python-multipart==0.0.12